VTCCA-926 move azure headers to api proxy

+++ app/models/APIIndividualAccount.scala
@@ -20,17 +20,9 @@ import java.time.Instant
 
 import play.api.libs.json.Json
 
-case class PersonData(
-                       identifyVerificationId: String,
-                       firstName: String,
-                       lastName: String,
-                       organisationId: Long,
-                       addressUnitId: Int,
-                       telephoneNumber: String,
-                       mobileNumber: Option[String],
-                       emailAddress: String,
-                       governmentGatewayExternalId: String,
-                       effectiveFrom: Instant)
+case class PersonData(identifyVerificationId: String, firstName: String, lastName: String,
+                      organisationId: Long, addressUnitId: Long, telephoneNumber: String, mobileNumber: Option[String],
+                      emailAddress: String, governmentGatewayExternalId: String, effectiveFrom: Instant)
 
 object PersonData {
   implicit val format = Json.format[PersonData]
@@ -45,7 +37,7 @@ object APIIndividualAccount {
 case class APIIndividualAccountForOrganisation(identifyVerificationId: String,
                                                firstName: String,
                                                lastName: String,
-                                               addressUnitId: Int,
+                                               addressUnitId: Long,
                                                telephoneNumber: String,
                                                mobileNumber: Option[String],
                                                emailAddress: String,
